config: refactored convert logic into config/types
As more sugar is added, the config/convert.go file was poised to
continually grow. This commit splits all conversion logic for each
config type into a function in the given type's file. These functions
are added to a list of converters at the beginning of runtime, which is
then iterated over to perform all conversions when the convert function
is called.
