+++ config/types/filesystems.go
@@ -14,6 +14,11 @@
 
 package types
 
+import (
+	ignTypes "github.com/coreos/ignition/config/v2_0/types"
+	"github.com/coreos/ignition/config/validate/report"
+)
+
 type Filesystem struct {
 	Name  string `yaml:"name"`
 	Mount *Mount `yaml:"mount"`
@@ -29,4 +34,38 @@ type Mount struct {
 type Create struct {
 	Force   bool     `yaml:"force"`
 	Options []string `yaml:"options"`
+}
+
+func init() {
+	register2_0(func(in Config, out ignTypes.Config) (ignTypes.Config, report.Report) {
+		for _, filesystem := range in.Storage.Filesystems {
+			newFilesystem := ignTypes.Filesystem{
+				Name: filesystem.Name,
+				Path: func(p ignTypes.Path) *ignTypes.Path {
+					if p == "" {
+						return nil
+					}
+
+					return &p
+				}(ignTypes.Path(filesystem.Path)),
+			}
+
+			if filesystem.Mount != nil {
+				newFilesystem.Mount = &ignTypes.FilesystemMount{
+					Device: ignTypes.Path(filesystem.Mount.Device),
+					Format: ignTypes.FilesystemFormat(filesystem.Mount.Format),
+				}
+
+				if filesystem.Mount.Create != nil {
+					newFilesystem.Mount.Create = &ignTypes.FilesystemCreate{
+						Force:   filesystem.Mount.Create.Force,
+						Options: ignTypes.MkfsOptions(filesystem.Mount.Create.Options),
+					}
+				}
+			}
+
+			out.Storage.Filesystems = append(out.Storage.Filesystems, newFilesystem)
+		}
+		return out, report.Report{}
+	})
 }
