+++ config/types/networkd.go
@@ -14,6 +14,11 @@
 
 package types
 
+import (
+	ignTypes "github.com/coreos/ignition/config/v2_0/types"
+	"github.com/coreos/ignition/config/validate/report"
+)
+
 type Networkd struct {
 	Units []NetworkdUnit `yaml:"units"`
 }
@@ -21,4 +26,16 @@ type Networkd struct {
 type NetworkdUnit struct {
 	Name     string `yaml:"name"`
 	Contents string `yaml:"contents"`
+}
+
+func init() {
+	register2_0(func(in Config, out ignTypes.Config) (ignTypes.Config, report.Report) {
+		for _, unit := range in.Networkd.Units {
+			out.Networkd.Units = append(out.Networkd.Units, ignTypes.NetworkdUnit{
+				Name:     ignTypes.NetworkdUnitName(unit.Name),
+				Contents: unit.Contents,
+			})
+		}
+		return out, report.Report{}
+	})
 }
